+++ source_code/watering_system_k60/Sources/watering_system.h
@@ -20,25 +20,25 @@
 #define EVENT_MASK 1
 
 typedef struct {
-	uint32_t total_time;
-	uint32_t passed_time;
+	uint64_t total_time;
+	uint64_t passed_time;
 }watering_system_progress, *watering_system_progress_ptr;
 
 typedef struct {
-	uint32_t start_delay;
+	uint64_t start_delay;
 	uint32_t pumping_time;
 	uint32_t current_pump_time;
-	uint32_t time_between_watering;
+	uint64_t time_between_watering;
 	uint32_t watering_cycles;
-	uint32_t dry_time;
+	uint64_t dry_time;
 	pump_ptr pump;
 	LWEVENT_STRUCT pump_event;
 	watering_system_progress progress;
 } watering_system_params, *watering_system_params_ptr;
 
-void watering_system(uint32_t start_delay, uint32_t time_between_watering, uint32_t watering_cycles, uint32_t pumping_time, uint32_t dry_time);
+void watering_system(uint64_t start_delay, uint64_t time_between_watering, uint32_t watering_cycles, uint32_t pumping_time, uint64_t dry_time);
 void watering_pump_task(uint32_t);
-void watering_system_update(uint32_t start_delay, uint32_t time_between_watering, uint32_t watering_cycles, uint32_t pumping_time, uint32_t dry_time);
+void watering_system_update(uint64_t start_delay, uint64_t time_between_watering, uint32_t watering_cycles, uint32_t pumping_time, uint64_t dry_time);
 void watering_system_pump_water(uint32_t duration);
 watering_system_progress_ptr watering_system_get_progress(void);
 watering_system_params_ptr watering_system_get_params(void);
